# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Rest
  description: Приложения для управления пользователями и картами с JWT авторизацией
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - JWT: [ ]
paths:
...
  /api/cards:
    get:
      tags:
        - Управление картами
      summary: Получить все карты (только для ADMIN)
      operationId: getAllCards
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            minimum: 1
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseCardResponse"
        "403":
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Управление картами
      summary: Создать новую карту
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/cards/{cardId}/status:
    patch:
      tags:
        - Управление картами
      summary: "Изменить статус карты, для админа"
      operationId: updateCardStatus
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
              enum:
                - ACTIVE
                - BLOCKED
                - EXPIRED
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cards/{cardId}/block:
    patch:
      tags:
        - Управление картами
      summary: "Заблокировать карту, для пользователя"
      operationId: blockCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "400":
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/transactions/card/{cardId}:
    get:
      tags:
        - Управление транзакциями
      summary: Получить историю транзакций по карте
      operationId: getCardTransactions
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseTransactionResponse"
        "403":
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cards/{cardId}:
    get:
      tags:
        - Управление картами
      summary: Получить карту по ID
      operationId: getCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "403":
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Управление картами
      summary: Удалить карту
      operationId: deleteCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
        "403":
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cards/user/{userId}:
    get:
      tags:
        - Управление картами
      summary: Получить все карты пользователя
      operationId: getUserCards
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            minimum: 1
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseCardResponse"
        "403":
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    TransactionResponse:
      type: object
      description: Ответ с информацией о транзакции
      properties:
        id:
          type: integer
          format: int64
          description: ID транзакции
          example: 1
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          description: Сумма транзакции
          example: 1000.5
        transactionDate:
          type: string
          format: date-time
          description: Дата и время транзакции
          example: 2024-01-15T14:30:00
        status:
          type: string
          description: Статус транзакции
          enum:
            - PENDING
            - COMPLETED
            - FAILED
            - CANCELLED
          example: COMPLETED
        description:
          type: string
          description: Описание транзакции
          example: Перевод за услуги
    CardCreateRequest:
      type: object
      description: Запрос на создание карты
      properties:
        cardHolder:
          type: string
          description: Держатель карты
          example: IVAN IVANOV
          maxLength: 100
          minLength: 2
          pattern: "^[A-Z\\s]+$"
      required:
        - cardHolder
    CardResponse:
      type: object
      description: Ответ с информацией о карте
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        maskedNumber:
          type: string
          description: Маскированный номер карты
          example: 123456******5678
        cardHolder:
          type: string
          description: Держатель карты
          example: IVAN IVANOV
        expirationDate:
          type: string
          format: date
          description: Срок действия карты
          example: 2025-12-31
        status:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: Баланс карты
          example: 1500.75
...
    PageResponseTransactionResponse:
      type: object
      description: Страничный ответ с пагинацией
      properties:
        content:
          type: array
          description: Список элементов на текущей странице
          items:
            $ref: "#/components/schemas/TransactionResponse"
        currentPage:
          type: integer
          format: int32
          description: Текущая страница (начинается с 0)
          example: 0
        totalPages:
          type: integer
          format: int32
          description: Общее количество страниц
          example: 5
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
          example: 50
        pageSize:
          type: integer
          format: int32
          description: Размер страницы
          example: 10
    PageResponseCardResponse:
      type: object
      description: Страничный ответ с пагинацией
      properties:
        content:
          type: array
          description: Список элементов на текущей странице
          items:
            $ref: "#/components/schemas/CardResponse"
        currentPage:
          type: integer
          format: int32
          description: Текущая страница (начинается с 0)
          example: 0
        totalPages:
          type: integer
          format: int32
          description: Общее количество страниц
          example: 5
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
          example: 50
        pageSize:
          type: integer
          format: int32
          description: Размер страницы
          example: 10
    ErrorResponse:
      description: Ответ с информацией об ошибке
      properties:
        error:
          type: string
          description: Сообщение об ошибке
          example: Карта не найдена
  securitySchemes:
    JWT:
      type: http
      name: JWT
      in: header
      scheme: bearer
      bearerFormat: JWT